requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django==3.1.7
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
